Integrate readFromString, writeToString, Endian into Utils.ts
